@@ -1,6 +1,6 @@
 package com.github.gavvydizzle.playertags.commands;
 
-import com.github.gavvydizzle.playertags.tag.TagsManager;
+import com.github.gavvydizzle.playertags.gui.InventoryManager;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
 import org.bukkit.command.TabExecutor;
@@ -8,42 +8,28 @@ import org.bukkit.entity.Player;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 
 public class PlayerCommandManager implements TabExecutor {
 
-    private final TagsManager tagsManager;
+    private final InventoryManager inventoryManager;
 
-    public PlayerCommandManager(TagsManager tagsManager) {
-        this.tagsManager = tagsManager;
+    public PlayerCommandManager(InventoryManager inventoryManager) {
+        this.inventoryManager = inventoryManager;
     }
 
     @Override
     public boolean onCommand(@NotNull CommandSender sender, @NotNull Command command, @NotNull String s, @NotNull String[] args) {
-        if (!(sender instanceof Player)) return true;
-        Player player = (Player) sender;
+        if (!(sender instanceof Player player)) return true;
 
-        if (args.length == 0) {
-            tagsManager.openTagsMenu(player, 1);
-        }
-        else {
-            int page;
-            try {
-                page = Integer.parseInt(args[0]);
-            } catch (NumberFormatException e) {
-                page = 1;
-            }
-
-            tagsManager.openTagsMenu(player, page);
-        }
-
+        inventoryManager.openTagSelectionMenu(player);
         return true;
     }
 
     @Nullable
     @Override
     public List<String> onTabComplete(@NotNull CommandSender sender, @NotNull Command command, @NotNull String s, @NotNull String[] args) {
-        return new ArrayList<>();
+        return Collections.emptyList();
     }
 }
